--- a/Src/DSInternals.Common.Test/SearchableDeviceKeyTester.cs
+++ b/Src/DSInternals.Common.Test/SearchableDeviceKeyTester.cs
@@ -38,10 +38,7 @@ namespace DSInternals.Common.Test
             Assert.AreEqual("cb69481e-8ff7-4039-93ec-0a2729a154a8", keyCredential.FidoKeyMaterial.AuthenticatorData.AttestedCredentialData.AaGuid.ToString());
 
             // Serialize the object again and compare with the original
-            string normalized = JsonSerializer.Serialize(
-                DsiJson.DeserializeLenient<JsonElement>(jsonData),
-                DsiJson.Options);
-            Assert.AreEqual(normalized, keyCredential.ToJson());
+            JsonAssert.AreEqual(jsonData, keyCredential.ToJson());
         }
 
         [TestMethod]
@@ -94,10 +91,7 @@ namespace DSInternals.Common.Test
             Assert.AreEqual(2, keyCredential.FidoKeyMaterial.AttestationCertificates.Count);
 
             // Serialize the object again and compare with the original
-            string normalized = JsonSerializer.Serialize(
-                DsiJson.DeserializeLenient<JsonElement>(jsonData),
-                DsiJson.Options);
-            Assert.AreEqual(normalized, keyCredential.ToJson());
+            JsonAssert.AreEqual(jsonData, keyCredential.ToJson());
         }
 
         [TestMethod]
@@ -125,10 +119,7 @@ namespace DSInternals.Common.Test
             Assert.AreEqual("cbad3c94-b480-4fa6-9187-ff1ed42c4479", parsedKey.DeviceId.Value.ToString().ToLowerInvariant());
 
             // Serialize the object again and compare with the original
-            string normalized = JsonSerializer.Serialize(
-                DsiJson.DeserializeLenient<JsonElement>(jsonData),
-                DsiJson.Options);
-            Assert.AreEqual(normalized, parsedKey.ToJson());
+            JsonAssert.AreEqual(jsonData, parsedKey.ToJson());
 
             // Re-generate the identifier and check that it matches the value in AAD.
             var generatedKey = new KeyCredential(
@@ -140,7 +131,7 @@ namespace DSInternals.Common.Test
             Assert.AreEqual(parsedKey.Identifier, generatedKey.Identifier);
 
             // Serialize the generated object and compare with the original
-            Assert.AreEqual(normalized, generatedKey.ToJson());
+            JsonAssert.AreEqual(jsonData, generatedKey.ToJson());
         }
 
         [TestMethod]