Here's a second version after updates that required some changes to tests as the differences of semantic JSON parsing

```

BenchmarkDotNet v0.15.2, Windows 11 (10.0.26100.4652/24H2/2024Update/HudsonValley)
Unknown processor
.NET SDK 9.0.304
  [Host]     : .NET 8.0.19 (8.0.1925.36514), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.19 (8.0.1925.36514), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI


```
| Method            | Mean       | Error    | StdDev   | Gen0   | Gen1   | Allocated |
|------------------ |-----------:|---------:|---------:|-------:|-------:|----------:|
| STJ_String_Small  |   283.3 ns |  3.32 ns |  3.10 ns | 0.0448 |      - |     752 B |
| STJ_Bytes_Small   |   243.5 ns |  2.57 ns |  2.15 ns | 0.0448 |      - |     752 B |
| STJ_String_Medium |   503.9 ns |  4.85 ns |  4.05 ns | 0.2584 | 0.0038 |    4336 B |
| STJ_Bytes_Medium  |   410.1 ns |  7.92 ns |  8.13 ns | 0.2589 | 0.0038 |    4336 B |
| STJ_String_Large  | 3,994.9 ns | 38.03 ns | 35.57 ns | 3.9215 | 0.7782 |   65776 B |
| STJ_Bytes_Large   | 2,881.5 ns |  9.88 ns |  7.71 ns | 3.9215 | 0.7820 |   65776 B |


## Benchmark Comparison – Original vs New vs Final

### ⏱️ Throughput (ns)

| Scenario            | Original | New (pre-latest) | Final | Δ New vs Orig | Δ Final vs Orig | Δ Final vs New |
|---------------------|---------:|-----------------:|------:|--------------:|----------------:|---------------:|
| **STJ_String_Small**  |   279.8 ns |   287.1 ns | **283.3 ns** | −2.6% | −1.2% | +1.3% |
| **STJ_Bytes_Small**   |   342.9 ns |   247.2 ns | **243.5 ns** | +27.9% | **+29.0%** | +1.5% |
| **STJ_String_Medium** |   631.9 ns | **499.8 ns** | 503.9 ns | **+20.9%** | **+20.3%** | −0.8% |
| **STJ_Bytes_Medium**  |   777.3 ns |   512.3 ns | **410.1 ns** | +34.1% | **+47.2%** | +20.0% |
| **STJ_String_Large**  | 4,043.0 ns | 4,610.5 ns | **3,994.9 ns** | −14.0% | **+1.2%** | +13.4% |
| **STJ_Bytes_Large**   | 36,013.7 ns | 3,421.0 ns | **2,881.5 ns** | +90.5% | **+92.0%** | +15.8% |

- **Δ** columns are speedup vs the left baseline (positive = faster, negative = slower).
- **Bold** values indicate the best performer.

---

### 🧠 Allocations

| Scenario            | Original | New (pre-latest) | Final |
|---------------------|--------:|-----------------:|------:|
| **STJ_String_Small**  |    752 B |     752 B |   752 B |
| **STJ_Bytes_Small**   |  1,784 B | **752 B** | **752 B** |
| **STJ_String_Medium** |  4,336 B |   4,336 B | 4,336 B |
| **STJ_Bytes_Medium**  | 12,536 B | **4,336 B** | **4,336 B** |
| **STJ_String_Large**  | 65,776 B |  65,776 B | 65,776 B |
| **STJ_Bytes_Large**   |196,884 B | **65,776 B** | **65,776 B** |

**Reduction vs Original (Final):**
- Bytes-Small: −57.8%
- Bytes-Medium: −65.4%
- Bytes-Large: −66.6%

---

### ✅ Key Takeaways
- **Final version dominates on byte input** (the hot path in AD/LAPS):  
  - ~**92% faster** for large payloads.  
  - ~**2/3 fewer allocations**.  
- **String input regressions fixed or minimized:**  
  - Large strings are now **faster than original** (~+1%).  
  - Medium strings still ~20% faster than original (slightly slower than pre-latest, but within noise).  
  - Small strings within ~1% of original.

---

### 🚀 Final 
  - UTF-8 span fast path.  
  - UTF-16 decode to string.  
  - Lenient reader vs strict writer split.  
  - Bounded single-quote heuristic.