@MichaelGrafnetter 2025-08-17
I "jiggled" a little with dependency versions and it just started to work. But I am afraid that sooner or later, I will need to use the `IModuleAssemblyInitializer` code you posted, so thanks for that.