@MichaelGrafnetter 2025-08-17
Yes. Every time I want to demo Microsoft Graph PowerShell to someone, I end up with some assemblies not loading and having to update all modules. So even MSFT is unable to solve this. Do you remember when .NET Framework was introduced 20+ years ago? One of the key features was "solving DLL hell".